+++ src/tools/runk/libcontainer/src/lib.rs
@@ -3,8 +3,9 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-pub mod builder;
+pub mod activated_builder;
 pub mod cgroup;
 pub mod container;
+pub mod init_builder;
 pub mod status;
 pub mod utils;
