+++ src/tools/runk/src/commands/run.rs
@@ -4,7 +4,7 @@
 //
 
 use anyhow::Result;
-use libcontainer::{builder::InitContainerBuilder, container::ContainerAction};
+use libcontainer::{container::ContainerAction, init_builder::InitContainerBuilder};
 use liboci_cli::Run;
 use slog::{info, Logger};
 use std::path::Path;
